backend/requirements.txt
fastapi>=0.100.0,<1.0.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.9
alembic>=1.13.0
pytest>=8.0.0
httpx>=0.26.0
pillow>=10.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0